requirements.txt
# python -> 3.7

tensorflow-gpu==2.4
numpy
pandas
livelossplot
